fix: coin integer 타입으로 변경

+++ project/back/health/src/main/java/com/ssafy/health/domain/account/dto/response/UserDetailDto.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class UserDetailDto {
 
     private String nickname;
-    private Long coin;
+    private Integer coin;
 }
